refactor: 💡 Relocate logout button component
